feat(preset-arbitrary): make arbitrary variant standalone
