+++ alias.ts
@@ -20,6 +20,7 @@ export const alias: Record<string, string> = {
   '@unocss/preset-uno': r('./packages/preset-uno/src/'),
   '@unocss/preset-web-fonts': r('./packages/preset-web-fonts/src/'),
   '@unocss/preset-wind': r('./packages/preset-wind/src/'),
+  '@unocss/preset-arbitrary': r('./packages/preset-arbitrary/src/'),
   '@unocss/runtime': r('./packages/runtime/src/'),
   '@unocss/scope': r('./packages/scope/'),
   '@unocss/shared-common': r('./packages/shared-common/src/'),
