+++ packages/core/src/extractors/index.ts
@@ -1,2 +1,2 @@
-export { extractorSplit, arbitraryPropertyRE, quotedArbitraryValuesRE, defaultSplitRE, splitWithVariantGroupRE } from './split'
+export { extractorSplit, defaultSplitRE, splitWithVariantGroupRE } from './split'
 export { extractorSvelte } from './svelte'
