+++ packages/preset-mini/src/_variants/default.ts
@@ -5,7 +5,7 @@ import { variantBreakpoints } from './breakpoints'
 import { variantCombinators } from './combinators'
 import { variantColorsMediaOrClass } from './dark'
 import { variantLanguageDirections } from './directions'
-import { variantCssLayer, variantInternalLayer, variantScope, variantSelector, variantVariables } from './misc'
+import { variantCssLayer, variantInternalLayer, variantScope, variantSelector } from './misc'
 import { variantNegative } from './negative'
 import { variantImportant } from './important'
 import { variantCustomMedia, variantPrint } from './media'
@@ -40,5 +40,4 @@ export const variants = (options: PresetMiniOptions): Variant<Theme>[] => [
   variantScope,
 
   variantContainerQuery,
-  variantVariables,
 ]
