+++ packages/shared-common/src/index.ts
@@ -1,6 +1,7 @@
 import type { ExtractorContext, UnoGenerator } from '@unocss/core'
-import { arbitraryPropertyRE, escapeRegExp, isAttributifySelector, makeRegexClassGroup, quotedArbitraryValuesRE, splitWithVariantGroupRE } from '@unocss/core'
+import { escapeRegExp, isAttributifySelector, makeRegexClassGroup, splitWithVariantGroupRE } from '@unocss/core'
 import MagicString from 'magic-string'
+import { arbitraryPropertyRE, quotedArbitraryValuesRE } from '../../preset-arbitrary/src/split-arbitrary'
 
 // https://github.com/dsblv/string-replace-async/blob/main/index.js
 export function replaceAsync(string: string, searchValue: RegExp, replacer: (...args: string[]) => Promise<string>) {
