+++ packages/unocss/preset-arbitrary.d.ts
@@ -1,0 +1,2 @@
+export * from './dist/preset-arbitrary'
+export { default } from './dist/preset-arbitrary'
