+++ packages/unocss/src/preset-arbitrary.ts
@@ -1,0 +1,2 @@
+export * from '@unocss/preset-arbitrary'
+export { default } from '@unocss/preset-arbitrary'
