+++ test/assets/preset-wind-targets.ts
@@ -394,7 +394,6 @@ export const presetWindTargets: string[] = [
 
   // variants experimental
   '@hover-text-red',
-  '@hover:[[open]_&]:text-blue',
 
   // variants - multiple parents
   '@dark:contrast-more:p-10',
